requirements.txt
asgiref==3.8.1
autopep8==1.6.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework-camel-case==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
toml==0.10.2
typing_extensions==4.11.0
